# Do not build View watcher until the first `updated?` check