--- a/actionview/lib/action_view/railtie.rb
+++ b/actionview/lib/action_view/railtie.rb
@@ -116,7 +116,6 @@ module ActionView
         view_reloader = ActionView::CacheExpiry::ViewReloader.new(watcher: app.config.file_watcher)
 
         app.reloaders << view_reloader
-        view_reloader.execute
         app.reloader.to_run do
           require_unload_lock!
           view_reloader.execute