@casperisfine 2024-03-18
Makes sense to me to not constantly re-do the same thing.